:recycle: Refactored game logic and animations
