+++ app/src/components/PropertySlide.css
@@ -1,8 +1,3 @@
-.property-divider {
-    height: 100%;
-    width: 6px;
-}
-
 .property1 {
     position: absolute;
 }
